Refactor ss:clean-temp command

+++ app/Console/Commands/CleanUpTempFolder.php
@@ -5,6 +5,9 @@ namespace App\Console\Commands;
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\Storage;
 
+use function Laravel\Prompts\confirm;
+use function Laravel\Prompts\info;
+
 class CleanUpTempFolder extends Command
 {
     /**
@@ -12,28 +15,35 @@ class CleanUpTempFolder extends Command
      *
      * @var string
      */
-    protected $signature = 'ss:clean-temp';
+    protected $signature = 'ss:clean-temp {--force}';
 
     /**
      * The console command description.
      *
      * @var string
      */
-    protected $description = 'Command description';
+    protected $description = 'Delete all files in temp folder';
 
     /**
      * Execute the console command.
      */
     public function handle()
     {
-        $disk = Storage::disk('tmp');
-
-        $files = $disk->files();
-
-        foreach ($files as $file) {
-            $disk->delete($file);
+        if (!$this->option('force')) {
+            $this->alert('Make sure no Slips are being queued, this will delete the temporary files.');
         }
 
-        $this->info('Temp folder cleaned up');
+        if ($this->option('force') || confirm('Do you really want to continue?', false)) {
+            $disk = Storage::disk('tmp');
+
+            $files = $disk->files();
+
+            foreach ($files as $file) {
+                $disk->delete($file);
+            }
+            info('Temp folder cleaned up');
+        }
+
+
     }
 }
